refactor `emotion` module
