+++ src/audio_generators.py
@@ -1,21 +1,27 @@
 import asyncio
+import json
 import os
 import random
-import re
+import typing as t
 from pathlib import Path
 from uuid import uuid4
 
-from langchain_community.callbacks import get_openai_callback
+import openai
+from pydantic import BaseModel
 from pydub import AudioSegment
+from requests import HTTPError
 
-from src.config import ELEVENLABS_MAX_PARALLEL, OPENAI_MAX_PARALLEL, logger
-from src.emotions.generation import EffectGeneratorAsync, TextPreparationForTTSTaskOutput
-from src.emotions.utils import add_overlay_for_audio
-from src.lc_callbacks import LCMessageLoggerAsync
+from src.config import ELEVENLABS_MAX_PARALLEL, OPENAI_API_KEY, OPENAI_MAX_PARALLEL, logger
 from src.schemas import SoundEffectsParams
 from src.text_split_chain import SplitTextOutput
 from src.tts import sound_generation_consumed, tts_astream_consumed
-from src.utils import consume_aiter
+from src.utils import add_overlay_for_audio, auto_retry, get_audio_duration
+
+from .prompts import (
+    SOUND_EFFECT_GENERATION,
+    SOUND_EFFECT_GENERATION_WITHOUT_DURATION_PREDICTION,
+    TEXT_MODIFICATION_WITH_SSML,
+)
 
 
 class AudioGeneratorSimple:
@@ -52,6 +58,89 @@ class AudioGeneratorSimple:
                     ab.write(chunk)
 
         return audio_combined_fp
+
+
+class TextPreparationForTTSTaskOutput(BaseModel):
+    task: str
+    output: t.Any
+
+
+class EffectGeneratorAsync:
+    def __init__(self, predict_duration: bool, model_type: str = "gpt-4o"):
+        self.client = openai.AsyncOpenAI(api_key=OPENAI_API_KEY)
+        self.sound_effect_prompt = (
+            SOUND_EFFECT_GENERATION
+            if predict_duration
+            else SOUND_EFFECT_GENERATION_WITHOUT_DURATION_PREDICTION
+        )
+        self.text_modification_prompt = TEXT_MODIFICATION_WITH_SSML
+        self.model_type = model_type
+
+    @auto_retry
+    async def generate_text_for_sound_effect(self, text: str) -> dict:
+        """Asynchronous version to generate sound effect description."""
+        try:
+            completion = await self.client.chat.completions.create(
+                model=self.model_type,
+                messages=[
+                    {"role": "system", "content": self.sound_effect_prompt},
+                    {"role": "user", "content": text},
+                ],
+                response_format={"type": "json_object"},
+            )
+            # Extracting the output
+            chatgpt_output = completion.choices[0].message.content
+
+            # Parse and return JSON response
+            output_dict = json.loads(chatgpt_output)
+            logger.info("Successfully generated sound effect description: %s", output_dict)
+            return output_dict
+
+        except json.JSONDecodeError as e:
+            logger.error("Failed to parse the output text as JSON: %s", e)
+            raise RuntimeError(
+                f"Error: Failed to parse the output text as JSON.\nOutput: {chatgpt_output}"
+            )
+
+        except HTTPError as e:
+            logger.error("HTTP error occurred: %s", e)
+            raise RuntimeError(f"HTTP Error: {e}")
+
+        except Exception as e:
+            logger.error("Unexpected error occurred: %s", e)
+            raise RuntimeError(f"Unexpected Error: {e}")
+
+    @auto_retry
+    async def generate_parameters_for_sound_effect(
+        self, text: str, generated_audio_file: str | None = None
+    ) -> TextPreparationForTTSTaskOutput:
+        llm_output = await self.generate_text_for_sound_effect(text)
+        if generated_audio_file is not None:
+            llm_output["duration_seconds"] = get_audio_duration(generated_audio_file)
+            logger.info(
+                "Added duration_seconds to output based on generated audio file: %s",
+                generated_audio_file,
+            )
+        return TextPreparationForTTSTaskOutput(task="sound_effects", output=llm_output)
+
+    @auto_retry
+    async def add_emotion_to_text(self, text: str) -> TextPreparationForTTSTaskOutput:
+        completion = await self.client.chat.completions.create(
+            model=self.model_type,
+            messages=[
+                {"role": "system", "content": self.text_modification_prompt},
+                {"role": "user", "content": text},
+            ],
+            response_format={"type": "json_object"},
+        )
+        chatgpt_output = completion.choices[0].message.content
+        try:
+            output_dict = json.loads(chatgpt_output)
+            logger.info("Successfully modified text with emotional cues: %s", output_dict)
+            return TextPreparationForTTSTaskOutput(task="tts", output=output_dict)
+        except json.JSONDecodeError as e:
+            logger.error("Error in parsing the modified text: %s", e)
+            raise ValueError(f"error, output_text: {chatgpt_output}")
 
 
 class AudioGeneratorWithEffects:
@@ -111,8 +200,8 @@ class AudioGeneratorWithEffects:
                 outputs = await func(**params)
                 return outputs
 
-        task_emotion_code = "add_emotion"
-        task_effects_code = "add_effects"
+        task_code_tts = "tts"
+        task_code_sounde_effects = "sound_effects"
 
         tasks = []
 
@@ -138,10 +227,12 @@ class AudioGeneratorWithEffects:
         tasks_results: list[TextPreparationForTTSTaskOutput] = []
         tasks_results = await asyncio.gather(*tasks)
 
-        emotion_tasks_results = [x.output for x in tasks_results if x.task == task_emotion_code]
-        effects_tasks_results = [x.output for x in tasks_results if x.task == task_effects_code]
+        tts_tasks_results = [x.output for x in tasks_results if x.task == task_code_tts]
+        sound_effects_tasks_results = [
+            x.output for x in tasks_results if x.task == task_code_sounde_effects
+        ]
 
-        return emotion_tasks_results, effects_tasks_results
+        return tts_tasks_results, sound_effects_tasks_results
 
     async def _generate_tts_audio(
         self,
