+++ src/utils.py
@@ -1,8 +1,10 @@
 import wave
 from enum import StrEnum
+from pathlib import Path
 
 from httpx import Timeout
 from langchain_openai import ChatOpenAI
+from pydub import AudioSegment
 from tenacity import retry, stop_after_attempt, wait_random_exponential
 
 from src.config import logger
@@ -44,3 +46,45 @@ def write_raw_pcm_to_file(data: bytes, fp: str, n_channels: int, bytes_depth: in
         f.setsampwidth(bytes_depth)
         f.setframerate(sampling_rate)
         f.writeframes(data)
+
+
+def get_audio_duration(filepath: str) -> float:
+    """
+    Returns the duration of the audio file in seconds.
+
+    :param filepath: Path to the audio file.
+    :return: Duration of the audio file in seconds.
+    """
+    audio = AudioSegment.from_file(filepath)
+    # Convert milliseconds to seconds
+    duration_in_seconds = len(audio) / 1000
+    return round(duration_in_seconds, 1)
+
+
+def add_overlay_for_audio(
+    main_audio_filename: str,
+    sound_effect_filename: str,
+    output_filename: str | None = None,
+    cycling_effect: bool = True,
+    decrease_effect_volume: int = 0,
+) -> str:
+    try:
+        main_audio = AudioSegment.from_file(main_audio_filename)
+        effect_audio = AudioSegment.from_file(sound_effect_filename)
+    except Exception as e:
+        raise RuntimeError(f"Error loading audio files: {e}")
+
+    if cycling_effect:
+        while len(effect_audio) < len(main_audio):
+            effect_audio += effect_audio
+
+    effect_audio = effect_audio[: len(main_audio)]
+
+    if decrease_effect_volume > 0:
+        effect_audio = effect_audio - decrease_effect_volume
+    combined_audio = main_audio.overlay(effect_audio)
+
+    if output_filename is None:
+        output_filename = f"{Path(main_audio_filename).stem}_{Path(sound_effect_filename).stem}.wav"
+    combined_audio.export(output_filename, format="wav")
+    return output_filename
